+++ packages/react-ui/internal/CommonWrapper/utils/extractCommonProps.ts
@@ -1,5 +1,5 @@
-import type { NotCommonProps } from './types';
-import type { CommonPropsWithRootNodeRef } from './CommonWrapper';
+import type { NotCommonProps } from '../types';
+import type { CommonPropsWithRootNodeRef } from '../CommonWrapper';
 
 export const extractCommonProps = <P extends CommonPropsWithRootNodeRef>(
   props: P,